Create io utils module and move stuff there
